Move test for ipython repr of long typed lists from test_dispatcher to test_typedlist
